+++ BTCPayServer/Controllers/UIWalletsController.cs
@@ -383,10 +383,18 @@ namespace BTCPayServer.Controllers
         private async Task SendFreeMoney(Cheater cheater, WalletId walletId, DerivationSchemeSettings paymentMethod)
         {
             var c = this.ExplorerClientProvider.GetExplorerClient(walletId.CryptoCode);
+            var cashCow = cheater.GetCashCow(walletId.CryptoCode);
+#if ALTCOINS
+            if (walletId.CryptoCode == "LBTC")
+            {
+                await cashCow.SendCommandAsync("rescanblockchain");
+            }
+#endif
             var addresses = Enumerable.Range(0, 200).Select(_ => c.GetUnusedAsync(paymentMethod.AccountDerivation, DerivationFeature.Deposit, reserve: true)).ToArray();
+            
             await Task.WhenAll(addresses);
-            await cheater.CashCow.GenerateAsync(addresses.Length / 8);
-            var b = cheater.CashCow.PrepareBatch();
+            await cashCow.GenerateAsync(addresses.Length / 8);
+            var b = cashCow.PrepareBatch();
             Random r = new Random();
             List<Task<uint256>> sending = new List<Task<uint256>>();
             foreach (var a in addresses)
@@ -394,7 +402,7 @@ namespace BTCPayServer.Controllers
                 sending.Add(b.SendToAddressAsync((await a).Address, Money.Coins(0.1m) + Money.Satoshis(r.Next(0, 90_000_000))));
             }
             await b.SendBatchAsync();
-            await cheater.CashCow.GenerateAsync(1);
+            await cashCow.GenerateAsync(1);
 
             var factory = ServiceProvider.GetRequiredService<NBXplorerConnectionFactory>();
 
@@ -1381,9 +1389,9 @@ namespace BTCPayServer.Controllers
             return Ok();
         }
 
-        [HttpGet("{walletId}/labels")]
+        [HttpGet("{walletId}/labels.json")]
         [IgnoreAntiforgeryToken]
-        public async Task<IActionResult> GetLabels(
+        public async Task<IActionResult> LabelsJson(
             [ModelBinder(typeof(WalletIdModelBinder))] WalletId walletId,
             bool excludeTypes,
             string? type = null,
@@ -1397,12 +1405,60 @@ namespace BTCPayServer.Controllers
                 : await WalletRepository.GetWalletLabels(walletObjectId);
             return Ok(labels
                 .Where(l => !excludeTypes || !WalletObjectData.Types.AllTypes.Contains(l.Label))
-                .Select(tuple => new
+                .Select(tuple => new WalletLabelModel
                 {
-                    label = tuple.Label,
-                    color = tuple.Color,
-                    textColor = ColorPalette.Default.TextColor(tuple.Color)
+                    Label = tuple.Label,
+                    Color = tuple.Color,
+                    TextColor = ColorPalette.Default.TextColor(tuple.Color)
                 }));
+        }
+
+        [HttpGet("{walletId}/labels")]
+        public async Task<IActionResult> WalletLabels(
+            [ModelBinder(typeof(WalletIdModelBinder))]
+            WalletId walletId)
+        {
+            if (walletId.StoreId == null)
+                return NotFound();
+
+            var labels = await WalletRepository.GetWalletLabels(walletId);
+
+            var vm = new WalletLabelsModel
+            {
+                WalletId = walletId,
+                Labels = labels
+                    .Where(l => !WalletObjectData.Types.AllTypes.Contains(l.Label))
+                    .Select(tuple => new WalletLabelModel
+                    {
+                        Label = tuple.Label,
+                        Color = tuple.Color,
+                        TextColor = ColorPalette.Default.TextColor(tuple.Color)
+                    })
+            };
+
+            return View(vm);
+        }
+
+        [HttpPost("{walletId}/labels/{id}/remove")]
+        public async Task<IActionResult> RemoveWalletLabel(
+            [ModelBinder(typeof(WalletIdModelBinder))]
+            WalletId walletId, string id)
+        {
+            if (walletId.StoreId == null)
+                return NotFound();
+
+            var labels = new[] { id };
+            ;
+            if (await WalletRepository.RemoveWalletLabels(walletId, labels))
+            {
+                TempData[WellKnownTempData.SuccessMessage] = "The label has been successfully removed.";
+            }
+            else
+            {
+                TempData[WellKnownTempData.ErrorMessage] = "The label could not be removed.";
+            }
+
+            return RedirectToAction(nameof(WalletLabels), new { walletId });
         }
 
         private string GetImage(PaymentMethodId paymentMethodId, BTCPayNetwork network)