+++ BTCPayServer/Services/Invoices/InvoiceEntity.cs
@@ -386,7 +386,7 @@ namespace BTCPayServer.Services.Invoices
         }
         public void UpdateTotals()
         {
-            Rates = new Dictionary<string, decimal>();
+            Rates = new Dictionary<string, decimal>(StringComparer.InvariantCultureIgnoreCase);
             foreach (var p in GetPaymentMethods())
             {
                 Rates.TryAdd(p.Currency, p.Rate);
@@ -935,6 +935,14 @@ namespace BTCPayServer.Services.Invoices
                 Status == InvoiceStatusLegacy.Invalid;
         }
 
+        public bool IsSettled()
+        {
+            return Status == InvoiceStatusLegacy.Confirmed ||
+                   Status == InvoiceStatusLegacy.Complete ||
+                   (Status == InvoiceStatusLegacy.Expired &&
+                    ExceptionStatus is InvoiceExceptionStatus.PaidLate or InvoiceExceptionStatus.PaidOver);
+        }
+
         public override int GetHashCode()
         {
             return HashCode.Combine(Status, ExceptionStatus);
@@ -970,7 +978,15 @@ namespace BTCPayServer.Services.Invoices
         }
         public override string ToString()
         {
-            return Status.ToModernStatus().ToString() + (ExceptionStatus == InvoiceExceptionStatus.None ? string.Empty : $" ({ToString(ExceptionStatus)})");
+            return Status.ToModernStatus() + ExceptionStatus switch
+            {
+                InvoiceExceptionStatus.PaidOver => " (paid over)",
+                InvoiceExceptionStatus.PaidLate => " (paid late)",
+                InvoiceExceptionStatus.PaidPartial => " (paid partial)",
+                InvoiceExceptionStatus.Marked => " (marked)",
+                InvoiceExceptionStatus.Invalid => " (invalid)",
+                _ => ""
+            };
         }
     }